server/WhisperFlask/requirements.txt
# requirements.txt

flask
werkzeug
openai
python-dotenv
psycopg2-binary
SQLAlchemy
flask-cors
